make auth for user

+++ app/Http/Controllers/Admin/UserController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Role;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -70,6 +71,10 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        if(Gate::denies('edit-users'))
+        {
+            return redirect(route('users.index'));
+        }
         $roles = Role::all();
 
         return view("admin.users.edit")->with([
@@ -100,6 +105,10 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if(Gate::denies('delete-users'))
+        {
+            return redirect(route('users.index'));
+        }
         $user->roles()->detach();
         $user->delete();
